+++ docs/changes.rst
@@ -8,6 +8,7 @@ Released on ? 2017.
 
 * Add glossary entries using AJAX.
 * The logout now uses POST to avoid CSRF.
+* The API key token reset now uses POST to avoid CSRF.
 
 weblate 2.13.1
 --------------